Add custom extension to Day 9 to support sample
The default preamble size for Day 9 prohibits actually using the given
sample input through the actual solution. Since the input data does not
encode the preamble size that should be used, that must be provided
somehow. Rather than breaking the API, instead, a custom way to identify
this is added as a supported first line.

Basically, any first line that starts with a '#' is treated like a
preamble directive. If this is not provided, then 25 is still assumed.